+++ src/core/lib/memorelay-hub.ts
@@ -7,7 +7,7 @@
 
 import { WebSocketServer } from 'ws';
 
-import { UpgradeHandler } from '../../experimental/types/upgrade-handler';
+import { UpgradeHandler } from '../types/upgrade-handler';
 import { handleUpgrade } from './handle-upgrade';
 import { RelayEvent } from '../events/relay-event';
 import { RelayError } from '../errors/relay-error';
