+++ src/core/plugins/create-clients.ts
@@ -12,7 +12,7 @@ import { MemorelayClient } from '../lib/memorelay-client';
 import { WebSocketConnectedEvent } from '../events/web-socket-connected-event';
 import { DuplicateWebSocketError } from '../errors/duplicate-web-socket-error';
 import { MemorelayClientCreatedEvent } from '../events/memorelay-client-created-event';
-import { Handler } from '../../experimental/types/handler';
+import { Handler } from '../types/handler';
 import { MemorelayHub } from '../lib/memorelay-hub';
 
 /**
