+++ src/nip-0001-basic-protocol/plugins/broadcast-incoming-event-messages.test.ts
@@ -9,9 +9,9 @@ import { BroadcastEventMessageEvent } from '../events/broadcast-event-message-ev
 import { broadcastIncomingEventMessages } from './broadcast-incoming-event-messages';
 import { IncomingEventMessageEvent } from '../events/incoming-event-message-event';
 import { ClientEventMessage } from '../../lib/message-types';
-import { createSignedTestEvent } from '../../experimental/test/signed-test-event';
+import { createSignedTestEvent } from '../../test/signed-test-event';
 import { MemorelayClientDisconnectEvent } from '../../core/events/memorelay-client-disconnect-event';
-import { setupTestHubAndClient } from '../../experimental/test/setup-hub-and-memorelay-client';
+import { setupTestHubAndClient } from '../../test/setup-hub-and-memorelay-client';
 
 describe('broadcastIncomingEventMessages()', () => {
   describe('#IncomingEventMessageEvent', () => {
