+++ src/nip-0001-basic-protocol/plugins/drop-duplicate-incoming-event-messages.test.ts
@@ -7,8 +7,8 @@
 
 import { dropDuplicateIncomingEventMessages } from './drop-duplicate-incoming-event-messages';
 import { IncomingEventMessageEvent } from '../events/incoming-event-message-event';
-import { createSignedTestEvent } from '../../experimental/test/signed-test-event';
-import { setupTestHubAndClient } from '../../experimental/test/setup-hub-and-memorelay-client';
+import { createSignedTestEvent } from '../../test/signed-test-event';
+import { setupTestHubAndClient } from '../../test/setup-hub-and-memorelay-client';
 import { MemorelayClientDisconnectEvent } from '../../core/events/memorelay-client-disconnect-event';
 
 describe('dropDuplicateIncomingEventMessages()', () => {
