+++ src/nip-0001-basic-protocol/plugins/drop-duplicate-incoming-event-messages.ts
@@ -8,7 +8,7 @@
 import { MemorelayClientCreatedEvent } from '../../core/events/memorelay-client-created-event';
 import { MemorelayHub } from '../../core/lib/memorelay-hub';
 import { IncomingEventMessageEvent } from '../events/incoming-event-message-event';
-import { Handler } from '../../experimental/types/handler';
+import { Handler } from '../../core/types/handler';
 import { MemorelayClientDisconnectEvent } from '../../core/events/memorelay-client-disconnect-event';
 import { clearHandlers } from '../../core/lib/clear-handlers';
