+++ src/nip-0001-basic-protocol/plugins/generalize-outgoing-event-messages.ts
@@ -7,7 +7,7 @@
  */
 
 import { MemorelayClientCreatedEvent } from '../../core/events/memorelay-client-created-event';
-import { Handler } from '../../experimental/types/handler';
+import { Handler } from '../../core/types/handler';
 import { MemorelayClientDisconnectEvent } from '../../core/events/memorelay-client-disconnect-event';
 import { clearHandlers } from '../../core/lib/clear-handlers';
 import { MemorelayHub } from '../../core/lib/memorelay-hub';
