+++ src/nip-0001-basic-protocol/plugins/generalize-outgoing-notice-messages.test.ts
@@ -6,7 +6,7 @@
  */
 
 import { generalizeOutgoingNoticeMessages } from './generalize-outgoing-notice-messages';
-import { setupTestHubAndClient } from '../../experimental/test/setup-hub-and-memorelay-client';
+import { setupTestHubAndClient } from '../../test/setup-hub-and-memorelay-client';
 import { OutgoingGenericMessageEvent } from '../events/outgoing-generic-message-event';
 import { MemorelayClientDisconnectEvent } from '../../core/events/memorelay-client-disconnect-event';
 import { OutgoingNoticeMessageEvent } from '../events/outgoing-notice-message-event';
