+++ src/nip-0001-basic-protocol/plugins/index.test.ts
@@ -8,7 +8,7 @@
 import { basicProtocol } from '.';
 import { BasicEventEmitter } from '../../core/lib/basic-event-emitter';
 import { MemorelayHub } from '../../core/lib/memorelay-hub';
-import { setupTestHub } from '../../experimental/test/setup-hub-and-memorelay-client';
+import { setupTestHub } from '../../test/setup-hub-and-memorelay-client';
 
 describe('basicProtocol()', () => {
   it('should increase maxEventListeners of hub', () => {
