+++ src/nip-0001-basic-protocol/plugins/index.ts
@@ -7,7 +7,7 @@
 
 import { clearHandlers } from '../../core/lib/clear-handlers';
 import { MemorelayHub } from '../../core/lib/memorelay-hub';
-import { Handler } from '../../experimental/types/handler';
+import { Handler } from '../../core/types/handler';
 import { broadcastIncomingEventMessages } from './broadcast-incoming-event-messages';
 import { generalizeOutgoingEOSEMessages } from './generalize-outgoing-eose-messages';
 import { generalizeOutgoingEventMessages } from './generalize-outgoing-event-messages';
