+++ src/nip-0001-basic-protocol/plugins/reject-unrecognized-incoming-messages.ts
@@ -10,7 +10,7 @@ import { BadMessageError } from '../errors/bad-message-error';
 import { BasicEventEmitter } from '../../core/lib/basic-event-emitter';
 import { IncomingGenericMessageEvent } from '../events/incoming-generic-message-event';
 import { MemorelayClientCreatedEvent } from '../../core/events/memorelay-client-created-event';
-import { Handler } from '../../experimental/types/handler';
+import { Handler } from '../../core/types/handler';
 import { MemorelayClientDisconnectEvent } from '../../core/events/memorelay-client-disconnect-event';
 import { clearHandlers } from '../../core/lib/clear-handlers';
