+++ src/nip-0001-basic-protocol/plugins/validate-incoming-close-messages.ts
@@ -12,7 +12,7 @@ import { BasicEventEmitter } from '../../core/lib/basic-event-emitter';
 import { IncomingCloseMessageEvent } from '../events/incoming-close-message-event';
 import { IncomingGenericMessageEvent } from '../events/incoming-generic-message-event';
 import { MemorelayClientCreatedEvent } from '../../core/events/memorelay-client-created-event';
-import { Handler } from '../../experimental/types/handler';
+import { Handler } from '../../core/types/handler';
 import { MemorelayClientDisconnectEvent } from '../../core/events/memorelay-client-disconnect-event';
 import { clearHandlers } from '../../core/lib/clear-handlers';
