+++ src/nip-0001-basic-protocol/plugins/validate-incoming-event-messages.test.ts
@@ -7,9 +7,9 @@
 
 import { validateIncomingEventMessages } from './validate-incoming-event-messages';
 import { IncomingGenericMessageEvent } from '../events/incoming-generic-message-event';
-import { setupTestHubAndClient } from '../../experimental/test/setup-hub-and-memorelay-client';
+import { setupTestHubAndClient } from '../../test/setup-hub-and-memorelay-client';
 import { IncomingEventMessageEvent } from '../events/incoming-event-message-event';
-import { createSignedTestEvent } from '../../experimental/test/signed-test-event';
+import { createSignedTestEvent } from '../../test/signed-test-event';
 import { BadMessageError } from '../errors/bad-message-error';
 
 describe('validateIncomingEventMessages()', () => {
